pyproject.toml
[project]
name = "time-series-forecasting-cluserting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastdtw>=0.3.4",
    "fastparquet>=2024.11.0",
    "matplotlib>=3.10.1",
    "nbformat>=5.10.4",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "plotly>=6.0.1",
    "pyarrow>=19.0.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.4.0,<1.5.3",
    "yfinance>=0.2.55",
    "autogluon==1.2",
]
